+++ e2e-tests/tests.js
@@ -1316,6 +1316,11 @@ module.exports = {
 
 'use strict';
 
+// Global vars
+var gDoc = document,
+    gDocEl = gDoc.documentElement,
+    gWin = window;
+
 
 /**
  * Add a class to an element.
@@ -1486,22 +1491,62 @@ function jqLiteOne(element, type, callback, useCapture) {
 
 
 /**
+ * Get or set horizontal scroll position
+ * @param {Element} element - The DOM element
+ * @param {number} [value] - The scroll position
+ */
+function jqLiteScrollLeft(element, value) {
+  // get
+  if (value === undefined) {
+    if (element === gWin) {
+      var t = (gWin.pageXOffset || gDocEl.scrollLeft)
+      return t - (gDocEl.clientLeft || 0);
+    } else {
+      return element.scrollLeft;
+    }
+  }
+
+  // set
+  if (element === gWin) gWin.scrollTo(value, jqLiteScrollTop(gWin));
+  else element.scrollLeft = value;
+}
+
+
+/**
+ * Get or set vertical scroll position
+ * @param {Element} element - The DOM element
+ * @param {number} value - The scroll position
+ */
+function jqLiteScrollTop(element, value) {
+  //return _scrollPos(element, 'top', value);
+
+  // get
+  if (value === undefined) {
+    if (element === gWin) {
+      return (gWin.pageYOffset || gDocEl.scrollTop) - (gDocEl.clientTop || 0);
+    } else {
+      return element.scrollTop;
+    }
+  }
+
+  // set
+  if (element === gWin) gWin.scrollTo(jqLiteScrollLeft(gWin), value);
+  else element.scrollTop = value;
+}
+
+
+/**
  * Return object representing top/left offset and element height/width.
  * @param {Element} element - The DOM element.
  */
 function jqLiteOffset(element) {
-  var win = window,
-      docEl = document.documentElement,
-      rect = element.getBoundingClientRect(),
-      viewLeft,
-      viewTop;
-
-  viewLeft = (win.pageXOffset || docEl.scrollLeft) - (docEl.clientLeft || 0);
-  viewTop = (win.pageYOffset || docEl.scrollTop) - (docEl.clientTop || 0);
+  var rect = element.getBoundingClientRect(),
+      scrollTop = jqLiteScrollTop(gWin),
+      scrollLeft = jqLiteScrollLeft(gWin);
 
   return {
-    top: rect.top + viewTop,
-    left: rect.left + viewLeft,
+    top: rect.top + scrollTop,
+    left: rect.left + scrollLeft,
     height: rect.height,
     width: rect.width
   };
@@ -1658,7 +1703,13 @@ module.exports = {
   removeClass: jqLiteRemoveClass,
 
   /** Check JavaScript variable instance type */
-  type: jqLiteType
+  type: jqLiteType,
+
+  /** Get or set horizontal scroll position */
+  scrollLeft: jqLiteScrollLeft,
+
+  /** Get or set vertical scroll position */
+  scrollTop: jqLiteScrollTop
 };
 
 
@@ -1674,11 +1725,13 @@ module.exports = {
 var config = require('../config.js'),
     jqLite = require('./jqLite.js'),
     win = window,
-    doc = window.document,
+    doc = document,
     nodeInsertedCallbacks = [],
+    scrollLock = 0,
+    scrollLockPos,
+    scrollLockEl,
     head,
     _supportsPointerEvents;
-
 
 head = doc.head || doc.getElementsByTagName('head')[0] || doc.documentElement;
 
@@ -1826,6 +1879,41 @@ function dispatchEventFn(element, eventType, bubbles, cancelable, data) {
 
 
 /**
+ * Turn on window scroll lock.
+ */
+function enableScrollLockFn() {
+  // increment counter
+  scrollLock += 1
+
+  // add lock
+  if (scrollLock === 1) {
+    scrollLockPos = {left: jqLite.scrollLeft(win), top: jqLite.scrollTop(win)};
+    scrollLockEl = loadStyleFn('body{overflow:hidden!important;}');
+    win.scrollTo(scrollLockPos.left, scrollLockPos.top);
+  }
+}
+
+
+/**
+ * Turn off window scroll lock.
+ */
+function disableScrollLockFn() {
+  // ignore
+  if (scrollLock === 0) return;
+
+  // decrement counter
+  scrollLock -= 1
+
+  // remove lock 
+  if (scrollLock === 0) {
+    scrollLockEl.parentNode.removeChild(scrollLockEl);
+    win.scrollTo(scrollLockPos.left, scrollLockPos.top);
+    scrollLockEl = null;
+  }
+}
+
+
+/**
  * Define the module API
  */
 module.exports = {
@@ -1835,9 +1923,15 @@ module.exports = {
   /** Classnames object to string */
   classNames: classNamesFn,
 
+  /** Disable scroll lock */
+  disableScrollLock: disableScrollLockFn,
+
   /** Dispatch event */
   dispatchEvent: dispatchEventFn,
   
+  /** Enable scroll lock */
+  enableScrollLock: enableScrollLockFn,
+
   /** Log messages to the console when debug is turned on */
   log: logFn,
 
@@ -2054,6 +2148,52 @@ describe('js/lib/jqLite.js', function() {
     });
 
     // TODO: specify missing values API
+  });
+
+
+
+  // --------------------------------------------------------------------------
+  // SCROLL METHODS
+  // --------------------------------------------------------------------------
+  describe('scroll methods', function() {
+    var el;
+
+
+    beforeEach(function() {
+      el = document.createElement('div');
+      document.body.appendChild(el);
+    });
+    
+
+    afterEach(function() {
+      el.parentNode.removeChild(el);
+    });
+
+
+    it('should set scroll position', function() {
+      var innerEl = document.createElement('div');
+      el.appendChild(innerEl);
+
+      // add overflow
+      jqLite.css(innerEl, {
+        'height': '100px',
+        'width': '100px',
+        'background-color': 'red'
+      });
+
+      jqLite.css(el, {
+        'height': '50px',
+        'width': '50px',
+        'overflow': 'auto'
+      });
+
+      // scroll
+      jqLite.scrollTop(el, 25);
+      assert.equal(jqLite.scrollTop(el), 25);
+
+      jqLite.scrollLeft(el, 30);
+      assert.equal(jqLite.scrollLeft(el), 30);
+    });
   });
 });